[#27] Rename scripts/ to src/
