+++ src/lateralus.component.collection.js
