+++ src/lateralus.component.js
