+++ src/lateralus.component.model.js
