+++ src/lateralus.component.view.js
